v2.16.0 add： 背成语功能 背单词添加空格提醒 赛马道具赛模式可以用参数控制了
fix：
修复入群事件报错
修复区域禁言处理
修复文转图文件夹删除失败的问题
修复哔哩哔哩视频解析 api 报错
修复点歌在网易云未搜索到歌曲时的报错问题

+++ util/text2image.py
@@ -60,8 +60,10 @@ def delete_old_cache():
     r = 0
     for cache_file in cache_files:
         i += 1
-        if cache_file.stat().st_mtime < (
-            (datetime.now() - timedelta(days=14)).timestamp()
+        if (
+            cache_file.stat().st_mtime
+            < ((datetime.now() - timedelta(days=14)).timestamp())
+            and cache_file.is_file()
         ):
             cache_file.unlink()
             r += 1
